+++ grafana-plugin/src/components/AlertTemplates/AlertTemplatesForm.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { SelectableValue } from '@grafana/data';
-import { getLocationSrv } from '@grafana/runtime';
 import { Label, Button, HorizontalGroup, VerticalGroup, Select, LoadingPlaceholder } from '@grafana/ui';
 import { capitalCase } from 'change-case';
 import cn from 'classnames/bind';
@@ -19,6 +18,7 @@ import { AlertReceiveChannel } from 'models/alert_receive_channel/alert_receive_
 import { Alert } from 'models/alertgroup/alertgroup.types';
 import { makeRequest } from 'network';
 import { UserAction } from 'state/userAction';
+import LocationHelper from 'utils/LocationHelper';
 
 import styles from './AlertTemplatesForm.module.css';
 
@@ -161,9 +161,7 @@ const AlertTemplatesForm = (props: AlertTemplatesFormProps) => {
       ) : null}
     </HorizontalGroup>
   );
-  const handleGoToTemplateSettingsCllick = () => {
-    getLocationSrv().update({ partial: true, query: { tab: 'Autoresolve' } });
-  };
+  const handleGoToTemplateSettingsCllick = () => LocationHelper.update({ tab: 'Autoresolve' }, 'partial');
 
   return (
     <div className={cx('root')}>